+++ WebApp/client/src/EditEmployee.jsx
@@ -17,7 +17,7 @@ function EditEmployee() {
 
   useEffect(() => {
     // Check user authentication using Axios
-    Axios.get("http://localhost:3000/isUserAuth", {
+    Axios.get("http://localhost:3000/authenticate/isUserAuth", {
       headers: {
         "x-access-token": localStorage.getItem("token"),
       },
@@ -36,7 +36,7 @@ function EditEmployee() {
 
   useEffect(() => {
     // Fetch employee data based on id_to_edit when the component mounts
-    Axios.get(`http://localhost:3000/employeeDetailForHR/${id_to_edit}`)
+    Axios.get(`http://localhost:3000/viewEmployee/${id_to_edit}`)
       .then(response => {
         // Set the fetched data to the state
         setEmployeeData(response.data.employee);
