+++ WebApp/client/src/PageAdmin.jsx
@@ -10,7 +10,7 @@ function PageAdmin() {
 
   useEffect(() => {
     // Check user authentication using Axios
-    Axios.get("http://localhost:3000/isUserAuth", {
+    Axios.get("http://localhost:3000/authenticate/isUserAuth", {
       headers: {
         "x-access-token": localStorage.getItem("token"),
       },
@@ -43,7 +43,7 @@ function PageAdmin() {
   const perPage = 10; // Number of records per page
 
   useEffect(() => {
-    Axios.get('http://localhost:3000/pageadmin/employeeData')
+    Axios.get('http://localhost:3000/employee/employeeData')
       .then((response) => setEmployees(response.data))
       .catch((error) => console.log(error));
   }, []);
