+++ WebApp/client/src/ViewHR.jsx
@@ -10,7 +10,7 @@ function ViewHR() {
 
   useEffect(() => {
     // Check user authentication using Axios
-    Axios.get("http://localhost:3000/isUserAuth", {
+    Axios.get("http://localhost:3000/authenticate/isUserAuth", {
       headers: {
         "x-access-token": localStorage.getItem("token"),
       },
@@ -28,7 +28,7 @@ function ViewHR() {
   }, [navigate]);
 
   useEffect(() => {
-    Axios.get(`http://localhost:3000/viewhr/employeeDetailForHR/${id_to_view}`)
+    Axios.get(`http://localhost:3000/viewEmployee/${id_to_view}`)
     .then((response) => {
         setEmployeeData(response.data.employee);
         setContactNumbers(response.data.contact);
