+++ WebApp/server/routes/Report3.js
@@ -1,18 +1,6 @@
 const express = require("express");
 const router = express.Router();
 
-router.get("/customAttributes", (req, res) => {
-  const db = req.db; // Access the 'db' object from the request
-  db.query("SELECT * FROM Custom_Attribute_Definition", (err, result) => {
-    if (err) {
-      console.log(err);
-      res.status(500).send("Error fetching custom attributes");
-    } else {
-      res.status(200).send(result);
-    }
-  });
-});
-
 // Fetch employee data with custom fields from the database
 router.get("/employeeCustomAttributes", (req, res) => {
   const db = req.db; // Access the 'db' object from the request
